cleanup: move clHCA to coding to simplify
Also has changed a lot from the original so it's barely external now
