+++ src/Makefile.autotools.am
@@ -10,10 +10,8 @@ SUBDIRS = coding layout meta
 # sources/headers are updated automatically by ./bootstrap script (not all headers are needed though)
 libvgmstream_la_LDFLAGS = coding/libcoding.la layout/liblayout.la meta/libmeta.la
 libvgmstream_la_SOURCES = (auto-updated)
-libvgmstream_la_SOURCES += ../ext_libs/clHCA.c
 libvgmstream_la_LIBADD = -lm
 EXTRA_DIST = (auto-updated)
-EXTRA_DIST += ../ext_includes/clHCA.h
 
 AM_CFLAGS += -DVGM_USE_G7221
