+++ src/libvgmstream.vcxproj
@@ -83,7 +83,7 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="..\ext_includes\clHCA.h" />
+    <ClInclude Include="coding\hca_decoder_clhca.h" />
     <ClInclude Include="coding\mpeg_bitreader.h" />
     <ClInclude Include="coding\mpeg_decoder.h" />
     <ClInclude Include="coding\vorbis_bitreader.h" />
@@ -622,6 +622,7 @@
     <ClCompile Include="coding\g7221_decoder_aes.c" />
     <ClCompile Include="coding\g7221_decoder_lib.c" />
     <ClCompile Include="coding\hca_decoder.c" />
+    <ClCompile Include="coding\hca_decoder_clhca.c" />
     <ClCompile Include="coding\ima_decoder.c" />
     <ClCompile Include="coding\imuse_decoder.c" />
     <ClCompile Include="coding\l5_555_decoder.c" />
@@ -706,7 +707,6 @@
     <ClCompile Include="layout\blocked_xa.c" />
     <ClCompile Include="layout\blocked_xa_aiff.c" />
     <ClCompile Include="layout\blocked_xvas.c" />
-    <ClCompile Include="..\ext_libs\clHCA.c" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
